requirements.txt
robin_stocks
python-dotenv
datetime
matplotlib